pyproject.toml
[project]
name = "bookcommerce"
version = "0.0.1-alpha"
description = "E-commerce for books"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-debug-toolbar>=5.2.0",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "django-rename-app>=0.1.7",
    "django-stubs>=5.1.3",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "pillow>=11.2.1",
    "psycopg>=3.2.6",
    "pygraphviz>=1.14",
]
